requirements.txt
# Jesse A. Eisenbalm LinkedIn Automation
# Core dependencies

# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.0.0

# AI & ML
openai>=1.12.0
google-genai>=0.3.0
tiktoken>=0.5.0

# Scheduling
APScheduler>=3.10.0

# Database
# (Using SQLite - no additional packages needed)

# HTTP & API
requests>=2.31.0
httpx>=0.26.0
aiohttp>=3.9.0

# Image Processing
Pillow>=10.0.0

# Data Processing
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# Utilities
python-dateutil>=2.8.0

# Optional: Development
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# black>=24.0.0
# ruff>=0.1.0
